mv hlper fun to utils
